send the email out to each group
Please note that in order to send these emails out, I used smtplib to host and had to use my email and password to log in. You will need to change sender_email and sender_password to your own information for this to work. If you are using gmail, you may have to change some settings in "Less secure app access."